Harden compositor against hostile clients; fix memory-safety bugs
From an adversarial code audit:
- ipc.c: IPC_APP_FRAME validated the frame origin in 32-bit arithmetic, so a
  hostile fx/fy near UINT32_MAX wrapped the bounds check and aimed a heap
  out-of-bounds write. Check in 64-bit. (critical)
- console.c: console_flip_if_dirty clamped only the top of the dirty band, so a
  rect pasted below the framebuffer made (y1-y0) underflow into a huge OOB VRAM
  memcpy. Clamp both ends before the empty-range test. (critical)
- wayland.c: wl_shm_pool.create_buffer never validated offset/size against the
  pool, so a client could over-map the fd and SIGBUS the whole desktop; validate
  in 64-bit. registry.bind and set_title could read past the message on crafted
  lengths; bound both. (high/low)
- gui_window.c: terminal scrollback wrote into a 256-byte stack buffer with an
  unclamped capacity, smashing the stack on wide/hi-DPI terminals. Clamp. (high)
- input.c: NUL-terminate EVIOCGNAME/PHYS results before printing. (medium)
- mouse.c: guard a cursor-outline shift that underflowed for b>16. (low)
- heap.c: route sizes 4081..4095 to the multi-page path (header pushed them over
  a page and kmalloc spuriously returned NULL). (bare-metal; high)

+++ fifi/platform/linux/ipc.c
@@ -339,9 +339,11 @@ static void ipc_dispatch(ipc_client_t *c, uint32_t type,
             }
         }
 
-        /* Update native frame_buf */
+        /* Update native frame_buf. Bounds check in 64-bit so a hostile fx/fy
+         * near UINT32_MAX cannot wrap fx+fw to a small value that passes the
+         * test and then aims the memcpy far past the frame buffer. */
         if (!c->minimized && c->frame_buf &&
-            fx + fw <= c->frame_w && fy + fh <= c->frame_h) {
+            (uint64_t)fx + fw <= c->frame_w && (uint64_t)fy + fh <= c->frame_h) {
             const uint32_t *src = (const uint32_t *)(pld + 16);
             for (uint32_t row = 0; row < fh; row++)
                 memcpy(c->frame_buf + (fy + row) * c->frame_w + fx,
